ED-1228 adiciona migration para ciar coluna nome em servicos

+++ database/migrations/2023_09_20_182637_add_column_name_on_services.php
@@ -15,8 +15,12 @@ return new class extends Migration
 
     public function down()
     {
-        Schema::table('services', function (Blueprint $table) {
-            $table->dropColumn('name');
-        });
+        if (Schema::hasColumn('services', 'name'))
+        {
+            Schema::table('services', function (Blueprint $table) {
+                $table->dropColumn('name');
+            });
+        }
+
     }
 };
